% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{add_rules}
\alias{add_rules}
\title{Function called if rule for gene did not exist and thus needs to be added}
\usage{
add_rules(g_env, app_env)
}
\arguments{
\item{g_env}{R environment, global environment}

\item{app_env}{R environment, app enviroment}
}
\description{
Function called if rule for gene did not exist and thus needs to be added
}
