% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cell_type}
\alias{check_cell_type}
\title{Checking whether a valid cell type was supplied by the user}
\usage{
check_cell_type(ct, app_env)
}
\arguments{
\item{ct}{string, cell type input by user}

\item{app_env}{R environment, app enviroment}
}
\description{
Checking whether a valid cell type was supplied by the user
}
