% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_all}
\alias{classify_all}
\title{Generate nearest neighbor per sample}
\usage{
classify_all(annotations)
}
\arguments{
\item{annotations}{list created by lc_vis}
}
\description{
Generate nearest neighbor per sample
}
