% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_sample_info}
\alias{generate_sample_info}
\title{Generate sample info data frame}
\usage{
generate_sample_info(cells)
}
\arguments{
\item{cells}{tibble, containing barcodes and corresponding sample}
}
\description{
Generate sample info data frame
}
