% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_app_env}
\alias{initialize_app_env}
\title{Populate the app environment}
\usage{
initialize_app_env(app_env, cells, counts, pc_space, embedding, nn, k)
}
\description{
Populate the app environment
}
