% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{lc_vis}
\alias{lc_vis}
\title{Visualize and annotate single cell data}
\usage{
lc_vis(cells, counts, pc_space, embedding, nn, k = 50)
}
\arguments{
\item{cells}{data.frame with cell metadata}

\item{counts}{count matrix (cols = cells, genes = rows)}

\item{pc_space}{PCA coordinates}

\item{embedding}{coordinates for two-dimensional embedding}

\item{nn}{list containing nearest neighbor information}

\item{k}{number of neighbors considered for smoothing}
}
\description{
Visualize and annotate single cell data
}
