% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_nn}
\alias{run_nn}
\title{Generate nearest neighbor per sample}
\usage{
run_nn(cells, pc_space, k = 50, dim = 30)
}
\arguments{
\item{cells}{data.frame, containing cell metadata}

\item{pc_space}{matrix, containing PCA coordinates (cols = coordiantes, rows = cells)}

\item{k}{numeric, number of neighbors considered for smoothing (default k = 50)}

\item{dim, }{numeric, number of PC dimensions used to compute nearest neighbor graph}
}
\description{
Generate nearest neighbor per sample
}
