% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{udpate_new_class}
\alias{udpate_new_class}
\alias{update_new_class}
\title{Function called if new class is created in a sample}
\usage{
update_new_class(g_env, app_env)
}
\arguments{
\item{g_env}{R environment, global environment}

\item{app_env}{R environment, app enviroment}
}
\description{
Function called if new class is created in a sample
}
