% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_annotation}
\alias{update_annotation}
\title{Function called after rules was added/deleted to update the cell type annotation}
\usage{
update_annotation(g_env, app_env)
}
\arguments{
\item{g_env}{R environment, global environment}

\item{app_env}{R environment, app enviroment}
}
\description{
Function called after rules was added/deleted to update the cell type annotation
}
