% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_new_sample}
\alias{update_new_sample}
\title{Update upon selection of new sample}
\usage{
update_new_sample(app_env, smp)
}
\arguments{
\item{app_env}{R environment, app enviroment}

\item{smp}{string, selected sample}
}
\description{
Update upon selection of new sample
}
