% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_rule}
\alias{update_rule}
\title{Function called if rule for gene exists and thus needs to be updated}
\usage{
update_rule(g_env, app_env)
}
\arguments{
\item{g_env}{R environment, global environment}

\item{app_env}{R environment, app enviroment}
}
\description{
Function called if rule for gene exists and thus needs to be updated
}
