% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_rules}
\alias{get_all_rules}
\title{Function to get all rules in a given sample}
\usage{
get_all_rules(rules, sample_select, all_types, current_class)
}
\description{
Function to get all rules in a given sample
}
