% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_classification}
\alias{run_classification}
\title{Generate nearest neighbor per sample}
\usage{
run_classification(annotations, sample_select)
}
\arguments{
\item{annotations}{list, created by lc_vis}

\item{sample_select}{string, selected sample which must be in sample_info$sample}
}
\description{
Generate nearest neighbor per sample
}
