% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_nn}
\alias{run_nn}
\title{Generate nearest neighbor per sample}
\usage{
run_nn(cells, pc_space, k = 50, dim = 30)
}
\arguments{
\item{cells}{data.frame with cell metadata}

\item{pc_space}{PCA coordinates}

\item{k}{number of neighbors considered for smoothing}

\item{dim}{number of PC dimensions used for nearest neighbor graph}
}
\description{
Generate nearest neighbor per sample
}
