% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{smoothed_fraction}
\alias{smoothed_fraction}
\title{function to calculate the fraction smoothed over nearest neighbors}
\usage{
smoothed_fraction(counts, marker, selection, nn_subset, gamma)
}
\description{
function to calculate the fraction smoothed over nearest neighbors
}
