% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{smoothed_fraction_wrapper}
\alias{smoothed_fraction_wrapper}
\title{Wrapper calling smoothed_fraction function and adding the result to the app environment}
\usage{
smoothed_fraction_wrapper(app_env)
}
\arguments{
\item{app_env}{R environment, app enviroment}
}
\description{
Wrapper calling smoothed_fraction function and adding the result to the app environment
}
